python-dfu/requirements.txt
note-python
pyserial
python-periphery
